Refactor merge.py into separate components
